Intermediate Machine Learning
1)  Introduction
Review what you need for this course.
2)  Missing Values
Missing values happen. Be prepared for this common challenge in real datasets.
3)  Categorical Variables
There's a lot of non-numeric data out there. Here's how to use it for machine learning.
4)   Pipelines
A critical skill for deploying (and even testing) complex models with pre-processing.
5)  Cross-Validation
A better way to test your models.
6)  XGBoost
The most accurate modeling technique for structured data.
7)  Data Leakage
Find and fix this problem that ruins your model in subtle ways.